Correction to SQL connection string
In the documentation it is stated as two-word connected and pascal cased string, like so: IntegratedSecurity=SSPI.
Reference link: https://docs.microsoft.com/en-us/sql/database-engine/availability-groups/windows/listeners-client-connectivity-application-failover?view=sql-server-ver15#ConnectToPrimary

But i encountered an error when trying to deploy my replica set configuration, which turned out to be because of a missing space in-between the variable. So after changing it to: Integrated Security=SSPI, it deployed successfully. Which is also supported by MSDN examples:

https://docs.microsoft.com/en-us/dotnet/api/system.data.sqlclient.sqlconnectionstringbuilder?view=netframework-4.8#remarks

https://docs.microsoft.com/en-us/dotnet/api/system.data.sqlclient.sqlconnectionstringbuilder.integratedsecurity?view=netframework-4.8#System_Data_SqlClient_SqlConnectionStringBuilder_IntegratedSecurity