Proposed Content - Untrusted Constraints
I have added proposed content for this page regarding untrusted constraints.  I have found through my usage of the DMA that the destination database can be left with untrusted constraints after the migration.  I'm assuming constraints are disabled and enabled again after the movement.  Leaving the constraints in an untrusted state can impact execution plans and performance.  I wanted to propose adding this content to check for this before and after the migration, and to properly fix them if needed.