Update database-files-and-filegroups.md
The following statement in the document is inaccurate:

"Read/write data filegroups and log files cannot be placed on an NTFS compressed file system. Only read-only databases and read-only secondary filegroups can be put on an NTFS compressed file system."

It is NOT Supported - True. It is NOT Recommended - True. It is a type of a hack - True as well. But it can be done! I showed how to do it in a post I published in 2013 in my blog:
http://ariely.info/Blog/tabid/83/EntryId/118/shrink-SQL-database-during-restore.aspx

I did a small change in the statement so it will provide the same information but will be accurate (replacing "cannot with not supported" for example:

My new version: "Read/write data filegroups and log files are not supported on an NTFS compressed file system. Only read-only databases and read-only secondary filegroups allowed to be put on an NTFS compressed file system."