Update Header Title to Reflect Document Name
The Header Title of this document was still *"GO (Transact-SQL)"*. The document itself states *"GO is **not** a Transact-SQL statement"*, and the title of the document reflects this therefore the title header should also be reflective of this. The incorrect information is also displayed in things like Google Searches, which could cause confusing for users performing searches.