Multiple grammar & missing word updates to automatic-tuning.md
In addition, for line 58, should it read something like:
Additionally, automatic tuning enables you to fully automate this process and let the Database Engine fix any problems found related to the plan change.

In addition, for line 68:
How long is the performance monitored for before Automatic tuning chooses to accept the change?
For line 175, 
How long does Azure SQL DB (Automatic Tuning) monitor query performance for, before it decides to accept the change? You say it is Immediately reverted, but in other articles it says the reversion might take up to 72 hours, depending on the volume of queries executing.