Updating verbiage for (lack of) SQL FCI support
Currently we say this:
Do not install R Services on a failover cluster. The security mechanism used for isolating R processes is not compatible with a Windows Server failover cluster environment.

But our 2017 docs say this:
Installing Machine Learning Services isn't supported on a failover cluster in SQL Server 2017. It's supported with SQL Server 2019.

When we say: