Adding Prefix to the LOCATION parameter
The LOCATION parameter was missing information about prefix. I took the syntax from CREATE EXTERNAL DATA SOURCE and added it here. '<prefix>://<path>[:<port>]' . Also, added a link to the LOCATION information in the CREATE EXTERNAL DATA SOURCE so that we don't have to repeat it here. The examples were already correct and using the syntax.