Update sqlpackage release notes with known issue
We found an issue for sqlpackage 18.5 where importing a bacpac with sensitivity classification using the .NET Core sqlpackage fails because of this known issue in Microsoft.Data.SqlClient. The fix is currently in the preview release of Microsoft.Data.SqlClient, so DacFx will be updated once the fix is in the stable release.