Clarification
SELECT isn't included as DML, but most vendors and docs note that SELECT is DML. I also added minor clarification to DDL, as this page is hard to update, and this list feels incomplete in terms of items that alter data structures. Arguably there might be other items in here, so added a general disclaimer.