SQL2017 query_post_execution_plan_profile overhead
following QFE 13515392.
Based on Jayc input : "On SQL17 however, lightweight profiling is OFF by default and query_post_execution_plan_profile depends on lightweight profile.
When you turn on/off this XEvent, the server will turn on/off lightweight profiling.   Every time you change lightweight profiling, the server needs to invalidate the execution plan cache.  So it is really bad idea to change this status frequently.  This is what the customer is doing every couple of minutes.  It would be better to turn lightweight profiling ON with TF7412."