Update configure-the-query-governor-cost-limit-server-configuration-option.md
This documentation misleadingly infers that the query cost is a measure of execution time when in fact this value is really the execution time on "Nick's machine". I have updated this document to reflect and clarify that the setting is not a measure of seconds on it's running instance to avoid the scenario where someone is using it to prevent queries for running for X seconds. I have also underlined the fact that the cost is based on estimations, not actuals.