Update set-query-governor-cost-limit-transact-sql.md
Related to my previous PR #5308 for the server configuration setting. This session configuration documentation misleadingly infers that the query cost is a measure of execution time when in fact this value is really the execution time on "Nick's machine". I have updated this document to reflect and clarify that the setting is not a measure of seconds on it's running instance to avoid the scenario where someone is using it to prevent queries for running for X seconds. I have also underlined the fact that the cost is based on estimations, not actuals. It also does not cancel running queries (like the documentation suggests) but instead will not execute a query over the threshold. I have fixed these issues.