Update dbcc-traceon-trace-flags-transact-sql.md
Several customer encounter the database recovery very long period issue which the database contains large in-memory table, especially there is no performance bottleneck. The following KB provide the trace flag 9944 as fix. So, I want to let this trace flag visible on central location.  
FIX: In-Memory OLTP database takes a long time to recover in SQL Server 2017 (https://support.microsoft.com/en-us/help/4090789/in-memory-oltp-database-takes-a-long-time-to-recover-in-sql-server)