Update merge-transact-sql.md
invalid example using merge into an edge-table. The example code is not working in SQL Server 2017.
A valid example is required, but I have no idea which code could work. At least a warning should be until a valid code example will be provided.
the "ON MATCH (Person-(livesIn)->City" causes the error when used in the merge statement.