Provided the full script for copy between subscriptions
Customers open a lot of support requests for copy between subscriptions. The finding the SID part might not be something they are familiar with. I have at least handled 10 customers with the same ask and I have always provided this script. It is better to have it in the doc itself to avoid support requests.