(AzureCXP) MicrosoftDocs/azure-docs#61909
Updating the repository URL because it is no more maintained.

Updated Jun 2020: This project is not being actively maintained. Instead, Apache Spark Connector for SQL Server and Azure SQL is now available, with support for Python and R bindings, an easier-to use interface to bulk insert data, and many other improvements. We encourage you to actively evaluate and use the new connector.