Update integration-services-service-ssis-service.md
The proposed change to delete this section was implemented in the original DMC, looks like it came back again during the consolidation of DMCs.
Removed the section "By default the SSIS service does not support del… #5127

This page is about SSIS Windows Service (MsDtsSrvr.exe) which is used for managing packages which are deployed and stored under MSDB, File System & SSIS package store using the old legacy package deployment model. The content on this section (which is proposed for deletion) is not relevant to the page. This section talks about SSIS Execution Runtime process (ISSERVEREXEC.exe) which is not related to SSIS Windows Service (MsDtsSrvr.exe). ISSERVEREXEC.exe is used only when executing the packages stored under SSISDB catalog using the new project deployment model. The information about delegation is already added under https://docs.microsoft.com/en-us/sql/integration-services/catalog/ssis-catalog?view=sql-server-ver15#ssisdb-catalog-and-delegation-in-double-hop-scenarios.