Adding RS 2019 product version
RS 2019 has two different versions (setup version and product version) which makes it confusing for customers and CSS. I have filed a bug on this (https://powerbi.visualstudio.com/Reach/_workitems/edit/437226) but product team states that it is too expensive to change to have only one version. So, I'm updating the doc to show both.