Make it more clear that 'tedious' is not just a word
Found this page while working on a test project and I was confused. I got the joke after reading the entire thing but having some formatting in the file would help a bit to call out that this is a term (ie, package name) rather than just being...the word tedious. :)