Update alter-database-transact-sql-set-options.md
removed the following from the sql db moniker range as it doesn't apply to SQL DB
you can change the ALLOW_SNAPSHOT_ISOLATION settings for the master, model, msdb, and tempdb databases. The setting is kept every time the instance of the Database Engine is stopped and restarted if you change the setting for tempdb. If you change the setting for model, that setting becomes the default for any new databases that are created, except for tempdb.


-------
cc: @markingmyname