Detail re Schema Limitations in Synapse Analytics
I have added a little detail in two places indicating that the Preview feature for Azure Synapse Analytics will only work in the DBO schema. The Development team is aware of this and indicated that they would be updating the documentation. https://feedback.azure.com/forums/307516-azure-synapse-analytics/suggestions/18434083-rename-a-column-name-in-sqldw#{toggle_previous_statuses}