Update sys-dm-hadr-availability-group-states-transact-sql.md
In "Not the primary replica OR unable to communicate with the WSFC failover cluster." the "OR" means if either condition is true. So when it's not the primary replica, I would expect to see NULL, However, I'm seeing the primary replica name when selecting the column on the secondary replica. In this case, since it _can_ communicate with the WSFC, it must be getting the primary replica name from the cluster. Therefore I believe it's NULL when 1) it's not the primary replica AND 2) it can't get the primary replica name from the WSFC failover cluster because it's unable to communicate with it.