Update checksum-transact-sql.md
Added caution remarks for two guaranteed hash collision scenarios.  I provided examples that should illustrate that this is a distinct caution from the doc's exisiting warning that hash non-collision is not guaranteed.  The latter reflects the relatively high probability of collision in a 4 byte modulus.  The former reflects a flaw (or intentional hack from days of yore?) in the hash function itself.