Parameter for time_window_in_hours is lowercase L, not capital i
We had a support case where the customer tried to use dta.exe and was getting an “Unrecognized option '-I' found.” error when trying to use '-I' (capital i) for time_window_in_hours parameter. It is actually '-l' (lowercase L). I made changes for the top Syntax block as well as the time_window_in_hours section in Arguments. The dta.exe -? command makes that clear in the output.