Improved memory-related wording
The current wording is incorrect. It suggests that as soon as you start using sp_xml_preparedocument that you lose 1/8 of SQL Server's memory. This is demonstratably not true. That's a limit. In addition, a suggestion is added that the nodes() method might be a better alternative in many cases, and avoid this issue in the first place.