Update alter-database-transact-sql-set-options.md
Note there are many options that are not supported in SQL MI per this link:  https://docs.microsoft.com/en-us/azure/azure-sql/managed-instance/transact-sql-tsql-differences-sql-server#alter-database-statement  including RESTRICTED_USER and READ_ONLY.  There are several others so this document should be updated to accurately reflect those options we do not support.  Please consider reviewing the full list of what is not supported and including that here to reduce confusion for our customers.