Add missing installation source for LocalDB
1. Add file location for SqlLocalDB.msi installer that is included with the installation media for all editions except for Express Core. There's no reason for people with Standard, Enterprise, or Developer editions to get and use the Express installer in order to download LocalDB when they already have the LocalDB installer.

I have verified by checking the install media for SQL Server 2017 Standard and Developer editions as well as SQL Server 2019 Developer edition. Enterprise should be the same as Developer, right? I don't have the ability to check any Web edition, but I think it's safe to assume that it's included there as well.

2. Minor formatting improvements (mainly fixing a few instances of `SqlLocalDb` to be `SqlLocalDB` like the rest)