Update sql-server-vss-writer-backup-guide.md on SQL Writer service state
Removed inaccuracies stating that SQL Writer is disabled by default. It is _enabled_ by default (automatic start). 
It's interesting to see the exact opposite is -correctly- mentioned in the very next paragraph.
There's no step to be done to configure it, it will be running and ready for operations right after setup is complete.