Added warning about requirements
Added a warning about some basic requirements or items to watch out for when restoring master, either to another different server or as part of a disaster recovery exercise. The warning is about supportability when the master database is restored to an instance that may or may not have all the same features and functionality of the original.