Don't recommend DATALENGTH as replacement for LEN
`LEN` counts the number of characters, while `DATALENGTH` counts the number of bytes. At the least, this requires logic bound to the type of the string (e.g. `varchar` vs. `nvarchar`), but also doesn't account for surrogate pairs that are counted as one character but multiple "characters" worth of bytes. `DATALENGTH` cannot be used to get the correct number of characters in a string containing surrogate pairs.