Update databasepropertyex-transact-sql.md
I would like to show how useful is DATABASEPROPERTYEX to identify when a query is running on a read-only replica when using the read scale out feature on Azure SQL Database. At this time I only see on this page useful examples of this function for SQL Server on-premises only.