Update row-compression-implementation.md
The document in question states the following for the SMALLDATETIME entry:

Uses the integer data representation by using two 2-byte integers. The date takes 2 bytes. It is the number of days since 1/1/1901. This needs 2 bytes starting from 1902. Therefore, there is no savings after that point.

That is incorrect. The SMALLDATETIME datatype using 1/1/1900 as its base date. You can see that in the documentation (https://docs.microsoft.com/en-us/sql/t-sql/data-types/smalldatetime-transact-sql?view=sql-server-ver15) for the datatype but you can also prove it in SQL Server by executing the following snippet of code...

 SELECT v.Day#, SmallDT = CONVERT(SMALLDATETIME,v.Day#)
   FROM (VALUES (0),(1),(2))v(Day#)
;

You can also prove that 1/1/1901 is NOT the base date for SMALLDATETIME using the following snippity of code...

 SELECT DaysSinceBaseDate = CONVERT(INT,CONVERT(SMALLDATETIME,'1/1/1901'))
;

It would also be real handy if the ISO representation for dates were uses, so, the correct statement should read...

Uses the integer data representation by using two 2-byte integers. The date takes 2 bytes. It is the number of days since 1900-01-01. This needs 2 bytes starting from 1900-09-14. Therefore, there is no savings after that point.