Added a more reliable drop database if exists option
We've had a lengthy discussion in the MVP email list about how to reliably drop a database if it exists. Most methods had issues and simply do not work reliably. This is the best method that we've come up with, and it should be recorded in the documentation for others to use.