Clean up examples - SARGability & clause positions
Update the example to use proper date boundaries, `>=` and `<` logic, in the final example. This changes the query to use a seek on the example index, rather than a scan of it; which would be much more performant as the query is SARGable. 

Also changed the date ranges in both example queries, as proposed in the related [issue](https://github.com/MicrosoftDocs/sql-docs/issues/6634) so that it's relevant for more recent versions of the Adventure Works database. Though I prefer, personally, to use unambiguous date formats (`yyyyMMdd`, I have used `CONVERT` in the second query with an explicit style to be consistent with the prior query.

Finally, I move the clause `ProductID BETWEEN 700 and 800` to the `WHERE`. `ProductID` is in the table `SalesOrderDetail` so it doesn't make much sense being in the `ON` clause for the `JOIN` to the table `SalesOrderHeader`. Technically the date boundaries in the `ON` clause in the second example could be moved to the `WHERE` (not if it were a `LEFT JOIN`) but it's not as unintuitive in its current position.