Add and scalability to the description for dev ed.
I noticed that developer edition isn't listed in the tables below.  I suspect that's because it is the same as Enterprise with the exception of the legal licensing scenarios.  There's some question though if it has the same unlimited scaling possibilities that SQL Server Enterprise does because we recall previous editions of SQL Server Developer lacking that.  I'm not sure if my change is warranted in the documentation.  If it is not, can Microsoft clarify this point for us?