The Info column does not exists by default
I build this query for my clients to show them an example of what can be part of that Indo column. I share it with you to help other clients, and to have a all the code needed for the samples in the articles. For example, the code is needed for the index in the article.