Update sys-dm-db-task-space-usage-transact-sql.md
|From|To|Relationship|  
|----------|--------|------------------|  
|dm_db_task_space_usage.request_id|dm_exec_requests.request_id|Many-to-one|  

The combination of session_id + request_id in sys.dm_db_task_space_usage might not always be unique in the scenario when parallel thread is spun using the same request_id as per the attached screenshot. In this situation, you would need to filter the sys.dm_db_task_space_usage DMV with exec_context_id = 0 to show only the active row.