This change adds new articles under Azure Data Studio section for the new extensions we are introducing:
- Extension for Oracle. Allows users to connect and explore existing objects in the Oracle database.
- Database Schema Conversion Toolkit (Oracle to MS SQL). An extension that allows users to convert Oracle database object definitions to Microsoft SQL object definitions.
I introduced one page for the Oracle extension along with a quick start guide, similarly to how other database extensions are doing it.

For Database Schema Conversion Toolkit I added a sub-folder with the landing page and further drill down to the Oracle to MS SQL scenario. In the future we plan to have more extensions as part of the toolkit branding for different source and target pairs, for example Oracle to PostgreSQL. These will have their own sub-folders to cover specifics of that scenario.

Finally, there is a subfolder that contains documentation articles for the conversion messages (messages that can be produced during conversion to draw attention to certain aspects of conversion). These messages are identified by the "code", so the articles reflect that. Some of the messages are generic enough to be shared across different scenarios, while others might be specific to a certain source-to-target scenario. I didn't want to separate them, as I don't think it matters for the user. We do not expect these articles themselves to be entry points. Instead we will link to them from the extensions themselves, whenever a certain message with certain code is generated - there will be a link to the associated documentation article.