Update dml-triggers.md
On line 49, I'm confused as to whether the trigger actually gets to do the final insert itself, or if it just allows or denies it. `before inserting the record into the payroll table`, I feel like, implies that at the end of the trigger is the one doing the inserting. It should be changed to something like `before the insertion of the record happens/takes place/finishes/succeeds/goes through` if it doesn't do the insert itself. Am I wrong?