Update logins-create-manage.md
Wording of
To create contained users mapped to Azure AD identities, you must be logged in using an Azure AD account that is an administrator (or db_owner role member) in the database in Azure SQL Database. In SQL Managed Instance, a SQL login with `sysadmin` permissions can also create an Azure AD login or user.

Is very poor. That kinda suggest that only someone set as AD server admin can do such action which is incorrect. "Administrator" word should be used with caution