Update restore-database-options-page.md
Based on RFC 14743166 
request to change the following:

From:
Preserves the replication settings when restoring a published database to a server other than the server where the database was created. This option is relevant only if the database was replicated when the backup was created.
This option is available only with the Leave the database ready for use by rolling back the uncommitted transactions option (described later in this table), which is equivalent to restoring a backup with the RECOVERY option.
Selecting this option is equivalent to using the KEEP_REPLICATION option in a RESTORE statement.

To
Preserves the replication settings when doing a restore using a backup file of a published database to a server other than the server where the database was created. This option is relevant only if the database was replicated when the backup was created. This option is available only with the Leave the database ready for use by rolling back the uncommitted transactions option (described later in this table), which is equivalent to restoring a backup with the RECOVERY option.
Selecting this option is equivalent to using the KEEP_REPLICATION option in a RESTORE statement.

When doing a restore using backup file of a published database to the same server where the database was created and with same database name, keep_replication is not needed.

When running “restore with recovery” and not restoring from a backup file, add keep_replication switch to preserve replication settings.