Update hadr-cluster-best-practices.md
Customer brought this to my attention that we call out 20 seconds (20000 ms) for our relaxed monitoring, but then give T-SQL for 15 seconds.

My gut take on this is that common sense would win out here and you'd update the value based on what you feel is best, but that didn't happen on at least a case I've had. My opinion following this case is that this T-SQL example would be better if the value we provide is set to match the value we provide for relaxed monitoring so there's no confusion at all.