Update parameter-sensitivity-plan-optimization.md
Updating line 61 to remove the reference to trace flag 12619.  Changed the sentence from:
Trace flag 12619 can be used to integrate with the Query Store if Query Store is enabled for your database. The following examples turn on the Query Store for a database called my_psp_database and enables trace flag 12619 globally:
to
For additional insights into the PSP optimization feature, we recommend that Query Store integration also be enabled by ensuring that the Query Store is turned on.  The following example will turn on the Query Store for a database called my_psp_database: