Update grant-database-scoped-credential-transact-sql.md
Verified the syntax on Dedicated SQL Pools in Synapse.
Syntax fails in Dedicated SQL Pool and only the Serverless SQL Pools works for this syntax in Synapse. 

Since the document is marked as supported for "Azure Synapse Analytics", we have to specifically mention which part is supported. 
Modified the page with "GRANT Database Scoped Credential Permissions is only supported for serverless SQL pools in Azure Synapse Analytics."