Update sys.sql_modules regarding sp_rename
The sp_rename article warns that renaming certain types of objects will not update the related information in sys.sql_modules. It would be helpful to have this warning in the sys.sql_modules article itself. (Also in object_definition, but doing this PR in the public editor for speed, so just tackling this one for the moment.)