Update sql-agent-extension-automatic-registration-all-vms.md
Discussing with Abdullah Mamun on this subject - the job doesn't run daily. It runs monthly. 

With that in mind, the documentation should reflect this, and the line below:

Automatic registration can take up to 4 days to detect newly created SQL Server VMs.

I doubt we can say that it will only take 4 days to pick up a VM with SQL Server installed. The process could take up to 30 depending on when SQL is installed, and when the job runs. 

Currently working a case now where this is the focus of the customer's issue, and again - if the job doesn't run daily, we can't have this out in the ether.