Add scenario to LOADHISTORY section
This additional paragraph is added after a user ran the restore with LOADHISTORY command for a set of backup files in the incorrect order (e.g. full backup, tlog backup 2, tlog backup 3, tlog backup 1).  The restore sequence is generated based on the backup_set_id (SMO does not interrogate LSN), so the restore sequence was incorrect.