Adding "Applies to" and Note for non-support for GMSAs
CSS team continues to receive support tickets related to customers using GMSAs as a SSRS/PBIRS service account. In August 2021, we had confirmed with the product team that gMSAs have not been tested and therefore are not currently supported. PG has a workitem to test gMSAs but it is on the backlog - https://powerbi.visualstudio.com/Power%20BI/_workitems/edit/655228. Requesting that doc is updated to inform customers of non-support and help CSS guide customers to official documentation on this scenario.