Added connected workspace reference
Recent SR made it clear that the feature disparity for renaming a dedicated SQL pool in a Synapse workspace is not clearly documented.  Renaming cannot occur under a standard workspace (where the workspace is created first and then the dedicated SQL pool is created), but is supported in a connected workspace scenario.