Using a real table in the Indexed Views verboten feature list
The current table under the "The SELECT statement in the view definition must not contain the following Transact-SQL elements:" bullet  is not a real table without any logical ordering or grouping which makes it hard to read.

This PR changes it to 3 separate tables: to keep functions separate from language elements and source column restrictions, and the middle table now groups language elements by their parent `SELECT` sub-clause. I feel this is much more readable - and it adds possible suggestions for alternative approaches - and makes it clear that while `COUNT` cannot be used, we can still use `COUNT_BIG`, which isn't immediately clear when reading the article as-is.

Updated the guidance to be more straightforward and to avoid possible alternatives that are confounded by other limitations. Making limited recommendations in this context is not always productive. Simplified the table instead of organizing by clause, which is not as straightforward for some elements. Simplified some language.