Removing ambiguity with date data type and date parameter
The syntax and description as currently written suggests that the parameter data type is a date. That's then confusing when other date-related data types come into use. I think it's important to make it clear that the 2nd parameter is not just a date, and to avoid mixing (throughout the discussion), the concepts of that date parameter and the date data type.