Update sql-server-backup-to-url-best-practices-and-troubleshooting.md
I don't know that this is the best means of reminding readers that the trace flag referenced is only applicable for page blobs, and not block blobs, but the number of times lately I've had customers AND customer engineers/CSAs point to that trace flag and then be stunned when they get no additional logging is pretty high. 

Which is what I'd expect knowing that the trace flag in question will only work for page blobs, not block blobs - which is what more and more of our customers have switched into.