Update database-import-export-hang.md with auto-cancellation note
To protect the service from abuse, we are rolling out a change which will automatically cancel jobs after two days. Customers must use their own resources to import or export databases which take too long in the service, or correct issues which are causing the database to run for too long, such as by increasing the DTU or VCore allocation.