This documentation is inaccurate
Using the portal, ledger cannot be enabled after database creation - it even says so in the page on the portal. The other examples all show it being enabled at database creation time. The portal example should be updated to show checking the box during the database creation time, using the checkbox on the security tab. I put in the basic steps, but for some reason I'm unable to configure automatic digest storage without an error so that may need to be researched.
-------
cc: @VanMSFT