Update dbcc-traceon-trace-flags-transact-sql.md
Clarified 2312 and 9481. These don't enable the user to set a CE model (reads as there's some subsequent choice/action to be made). These actually immediately set the intended CE model.

Also clarified 2446, that somehow was changed to read as applicable to "SQL 2017 and SQL 2022 and later" instead of the norm "SQL 2017 and SQL 2019 and later" - where SQL22 is therefore included. Same as for example 2422, 1800, etc