Some more emphasis on Registry prerequisites
It's easy to get lost in this article and all the steps that need to be done carefully to successfully set up TDE EKM with Key vault. I'd like some more emphasis on the registry prerequisites so people don't overlook and need to troubleshoot error messages*. I was not sure how the article would benefit most. This contribution vs moving up the registry prerequisites (and related note) in the step. 
 
*Msg 33049, Level 16, State 2, Line 30
Key with name 'kv-ContosoEKMKeyVaultKey' does not exist in the provider or access is denied. Provider error code: 2058.  (Provider Error - No explanation is available, consult EKM Provider for details)
 
-------
cc: @Rupp29