Adapt growth value to context and give clarity on growth unit
In case of a file growth set in MB, the original query gives the data file growth size in units of 8-KB pages, which is not really very clear for most people. Plus, the file size is in MB, so it lacks a bit of consistency. This proposal will give the growth value in MB in this case instead of 8-KB pages, and will keep percentage or 0 in case of fixed size.