Updated Typo Word from "data" to "users"
Under Advance Security and Compliance->Security Features->Row-Level Security. Updated a typo word from "data" to "users"
Statement was: For example, ensuring that workers can access only the data rows that are pertinent to their department, or restricting a data access to only the relevant data.
Statement Changed to: For example, ensuring that workers can access only the data rows that are pertinent to their department, or restricting a data access to only the relevant users.