Update sp-addpublication-transact-sql.md
Make it more clear. 

This parameter pertains solely to the manner in which the log reader agent operates. The present explanation inadequately illustrates this and employs apparent fact as the characterization. Consider, if the value is "snapshot", would the statement "Publisher provides output of all log-based transactions" still hold true? Indeed!