Clarify point in time restore geo-redundant backup
Point in time restore can be done in another server in a paired region so long as geo-redundant backup is selected.  This should be corrected to the effect of suggested changes, *you can run a restore only on the same server* => this inaccuracy slows down the process of restoration from customers who were impacted by recent outages.