Update sql-server-and-client-encryption-summary.md
Only edited line 46. The sys.dm_exec_connections DMV actually doesn't have an is_encrypted column, but it does have an encrypt_option column with true/false values. This ultimately provides a user with the same information about whether a connection is encrypted or not so no further changes are needed.